+++ Resources/Locale/en-US/shuttles/console.ftl
@@ -4,8 +4,8 @@ shuttle-pilot-end = Stopped piloting
 shuttle-console-in-ftl = Currently in FTL
 shuttle-console-mass = Too large to FTL
 shuttle-console-prevent = You are unable to pilot this ship
-shuttle-console-no-powered-ftl-drive = No powered FTL drive detected
-shuttle-console-ftl-drive-unpowered = FTL drive detected but not powered
+shuttle-console-no-powered-ftl-drive = No FTL drive detected.
+shuttle-console-ftl-drive-unpowered = FTL drive detected but not powered.
 
 # NAV
